Resolved monitor callbacks
Saved a activelistener in handler which would be updated if there is a new/change in booking otherwise it will be null. On server side, it will always check if handler's activelistener is null. If it is not, it will construct message to inform clients in activelistener.

Also corrected the construction of datagram packet for response in server to use response byte[].